Trying to fix Messages.sh, only run extractrc when there really are .rc or .ui files.

+++ Messages.sh
@@ -1,6 +1,12 @@
 #! /usr/bin/env bash
 subdirs="interfaces kross language outputview project shell sublime util vcs veritas"
-$EXTRACTRC `find $subdirs -name \*.rc` >> rc.cpp || exit 11
-$EXTRACTRC `find $subdirs -name \*.ui` >> rc.cpp || exit 12
+rcfiles="`find $subdirs -name \*.rc`"
+uifiles="`find $subdirs -name \*.ui`"
+if [[ "$rcfiles" != "" ]] ; then
+    $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
+fi
+if [[ "$uifiles" != "" ]] ; then
+    $EXTRACTRC  >> rc.cpp || exit 12
+fi
 $XGETTEXT -kaliasLocal `find $subdirs -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdevplatform.pot
 rm -f rc.cpp
